# openapi.yaml
openapi: 3.0.0
info:
  title: BorrowIt API
  description: |
    Backend API for BorrowIt - A rental platform where users can rent and lend items.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@borrowit.com
servers:
  - url: http://localhost:3456
    description: Development server
  - url: http://localhost:3456/api
    description: Development server with API prefix
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        username:
          type: string
          example: johndoe
        email:
          type: string
          format: email
          example: john@example.com
        full_name:
          type: string
          example: John Doe
        phone:
          type: string
          example: "+1234567890"
        avatar_url:
          type: string
          format: uri
          example: https://example.com/avatar.jpg
        address:
          type: string
          example: "123 Main St, City"
        role:
          type: string
          enum: [customer, seller, admin]
          example: customer
        status:
          type: string
          enum: [active, suspended, pending]
          example: active
        rating:
          type: number
          format: float
          example: 4.5
        total_orders:
          type: integer
          example: 10
        email_verified:
          type: boolean
          example: true
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Product:
      type: object
      properties:
        product_id:
          type: integer
          example: 1
        seller_id:
          type: integer
          example: 2
        name:
          type: string
          example: "Canon EOS R5 Camera"
        description:
          type: string
          example: "Professional mirrorless camera in excellent condition"
        category:
          type: string
          enum: [Electronics, Clothes, Furniture, Sports, Books, Tools, Vehicles, Other]
          example: Electronics
        price_per_day:
          type: number
          format: float
          example: 50.00
        sale_percentage:
          type: integer
          minimum: 0
          maximum: 100
          example: 10
        images:
          type: array
          items:
            type: string
            format: uri
          example: ["https://example.com/image1.jpg", "https://example.com/image2.jpg"]
        location:
          type: string
          example: "New York, NY"
        status:
          type: string
          enum: [available, rented, unavailable]
          example: available
        condition:
          type: string
          enum: [new, like_new, good, fair, poor]
          example: like_new
        rating:
          type: number
          format: float
          example: 4.8
        total_reviews:
          type: integer
          example: 25
        total_rentals:
          type: integer
          example: 15
        min_rental_days:
          type: integer
          example: 1
        max_rental_days:
          type: integer
          example: 30
        deposit:
          type: number
          format: float
          example: 200.00
        product_average_rating:
          type: number
          format: float
          example: 4.5
        seller_username:
          type: string
          example: johndoe
        seller_name:
          type: string
          example: John Doe
        seller_avatar:
          type: string
          format: uri
        seller_rating:
          type: number
          format: float
          example: 4.7
        seller_total_orders:
          type: integer
          example: 20
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        cart_id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 2
        sku:
          type: string
          example: "SKU12345"
        quantity:
          type: integer
          example: 1
        unit_price:
          type: number
          format: float
          example: 50.00
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        total_price:
          type: number
          format: float
          example: 150.00
        rent_time:
          type: integer
          example: 3
    OrderItem:
      type: object
      properties:
        productId:
          type: string
          example: PRD-005
        name:
          type: string
          example: "Designer Coat"
        size:
          type: string
          example: Large
        color:
          type: string
          example: Grey
        rentalPeriod:
          type: string
          example: "5 days"
        quantity:
          type: integer
          example: 1
        price:
          type: number
          format: float
          example: 300
        imageUrl:
          type: string
          format: uri
          example: https://images.example.com/item.jpg
    OrderTimelineEvent:
      type: object
      properties:
        title:
          type: string
          example: Order Placed
        date:
          type: string
          example: "September 16, 2025 - 11:00 AM"
        completed:
          type: boolean
          example: true
        description:
          type: string
          example: "Your order has been confirmed and payment received."
    ShippingAddress:
      type: object
      properties:
        name: { type: string, example: "John Doe" }
        address: { type: string, example: "123 Fashion Street" }
        city: { type: string, example: "New York" }
        state: { type: string, example: "NY" }
        zipCode: { type: string, example: "10001" }
        phone: { type: string, example: "+1 (555) 123-4567" }
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 12
        orderId:
          type: string
          example: ORD-25-2
        customerId:
          type: integer
          example: 5
        sellerId:
          type: integer
          example: 2
        sellerName:
          type: string
          example: "Premium Fashion"
        buyerName:
          type: string
          example: "John Doe"
        status:
          type: string
          example: Ordered
        placedDate:
          type: string
          example: "September 16, 2025"
        subtotal:
          type: number
          format: float
          example: 300
        tax:
          type: number
          format: float
          example: 24
        totalAmount:
          type: number
          format: float
          example: 324
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/OrderTimelineEvent'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        receivingInfo:
          type: object
          description: Optional receiving schedule selected by customer.
        returnInfo:
          type: object
          description: Optional return schedule selected by customer.
        notes:
          type: string
          example: "Leave with concierge if not home"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required:
        - productId
        - totalAmount
      properties:
        productId:
          type: string
          example: PRD-005
        productName:
          type: string
          example: "Designer Coat"
        sellerId:
          type: integer
          example: 12
        sellerName:
          type: string
          example: "Premium Fashion"
        quantity:
          type: integer
          example: 1
        rentalPeriod:
          type: string
          example: "5 days"
        status:
          type: string
          example: ordered
        placedAt:
          type: string
          format: date-time
        unitPrice:
          type: number
          format: float
          example: 300
        subtotal:
          type: number
          format: float
          example: 300
        tax:
          type: number
          format: float
          example: 24
        totalAmount:
          type: number
          format: float
          example: 324
        imageUrl:
          type: string
          format: uri
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/OrderTimelineEvent'
        receivingInfo:
          type: object
          description: Optional receiving details supplied by the renter.
        returnInfo:
          type: object
          description: Optional return schedule supplied by the renter.
        notes:
          type: string
        items:
          description: Optional custom items array; defaults to a single item built from the payload.
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    ReviewExperience:
      type: object
      properties:
        fit:
          type: boolean
        quality:
          type: boolean
        easeOfUse:
          type: boolean
        style:
          type: boolean
        worthThePrice:
          type: boolean
      example:
        fit: true
        quality: true
        easeOfUse: false
        style: true
        worthThePrice: true
    Review:
      type: object
      properties:
        review_id:
          type: integer
          example: 42
        reviewer_id:
          type: integer
          example: 7
        product_id:
          type: integer
          example: 3
        order_id:
          type: string
          example: ORD-25-1
        satisfaction:
          type: string
          enum: [loved-it, liked-it, it-was-okay, not-great, terrible]
        satisfaction_score:
          type: number
          format: float
          example: 4.5
        experience:
          $ref: '#/components/schemas/ReviewExperience'
        highlights:
          type: string
        improvements:
          type: string
        photos:
          type: array
          items:
            type: string
            format: uri
        photos_count:
          type: integer
        submitted_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ReviewCreatePayload:
      type: object
      required:
        - productId
        - satisfaction
      properties:
        productId:
          type: integer
          example: 3
        orderId:
          type: string
          example: ORD-25-2
        satisfaction:
          type: string
          enum: [loved-it, liked-it, it-was-okay, not-great, terrible]
        experience:
          $ref: '#/components/schemas/ReviewExperience'
        highlights:
          type: string
          maxLength: 2000
        improvements:
          type: string
          maxLength: 2000
        photos:
          type: array
          items:
            type: string
            format: uri
...
    Report:
      type: object
      properties:
        report_id:
          type: integer
          example: 1
        reporting_user_id:
          type: integer
          example: 2
        reported_user_id:
          type: integer
          example: 3
        reporting_user_username:
          type: string
          example: reporter123
        reported_user_username:
          type: string
          example: reportedUser
        reason:
          type: string
          example: "Inappropriate behavior"
        reason_extended:
          type: string
          example: "Detailed description of the issue..."
        image:
          type: string
          format: uri
          nullable: true
        status:
          type: string
          enum: [pending, in_review, resolved, dismissed]
          example: pending
        action:
          type: string
          nullable: true
        unsuspend_date:
          type: string
          format: date-time
          nullable: true
        action_reason:
          type: string
          nullable: true
        mod_note:
          type: string
          nullable: true
        report_date:
          type: string
          format: date-time
        resolve_date:
          type: string
          format: date-time
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: string
          example: Detailed error information
    PaginationResponse:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        totalPages:
          type: integer
          example: 10
paths:
...
  /api/users:
    get:
      summary: Get all users
      description: Retrieve a paginated list of all users with optional filters
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Items per page
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, seller, admin]
          description: Filter by user role
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, pending]
          description: Filter by user status
        - name: search
          in: query
          schema:
            type: string
          description: Search by username, email, or full name
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      pagination:
                        $ref: '#/components/schemas/PaginationResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/products:
    get:
      summary: Get all products
      description: Retrieve a paginated list of all products with optional filters
      tags: [Products]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Items per page
        - name: category
          in: query
          schema:
            type: string
            enum: [Electronics, Clothes, Furniture, Sports, Books, Tools, Vehicles, Other]
          description: Filter by category
        - name: status
          in: query
          schema:
            type: string
            enum: [available, rented, unavailable]
          description: Filter by status
        - name: minPrice
          in: query
          schema:
            type: number
          description: Minimum price per day
        - name: maxPrice
          in: query
          schema:
            type: number
          description: Maximum price per day
        - name: location
          in: query
          schema:
            type: string
          description: Filter by location
        - name: search
          in: query
          schema:
            type: string
          description: Search in product name and description
        - name: sortBy
          in: query
          schema:
            type: string
            default: created_at
            enum: [created_at, price_per_day, rating, name]
          description: Sort field
        - name: order
          in: query
          schema:
            type: string
            default: DESC
            enum: [ASC, DESC]
          description: Sort order
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
                      pagination:
                        $ref: '#/components/schemas/PaginationResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      summary: Create new product
      description: Create a new product listing (Seller only)
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sellerId
                - name
                - category
                - pricePerDay
                - location
              properties:
                sellerId:
                  type: integer
                  example: 2
                name:
                  type: string
                  example: "Canon EOS R5 Camera"
                description:
                  type: string
                  example: "Professional mirrorless camera"
                category:
                  type: string
                  enum: [Electronics, Clothes, Furniture, Sports, Books, Tools, Vehicles, Other]
                  example: Electronics
                pricePerDay:
                  type: number
                  format: float
                  example: 50.00
                salePercentage:
                  type: integer
                  minimum: 0
                  maximum: 100
                  example: 10
                images:
                  type: array
                  items:
                    type: string
                  example: ["https://example.com/image1.jpg"]
                location:
                  type: string
                  example: "New York, NY"
                condition:
                  type: string
                  enum: [new, like_new, good, fair, poor]
                  example: like_new
                minRentalDays:
                  type: integer
                  example: 1
                maxRentalDays:
                  type: integer
                  example: 30
                deposit:
                  type: number
                  format: float
                  example: 200.00
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product created successfully
                  data:
                    $ref: '#/components/schemas/Product'
        '404':
          description: Seller not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/cart:
    post:
      summary: Add an item to a cart
      description: Adds a new product to the specified cart.
      tags: [Cart]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - quantity
                - rentTime
              properties:
                productId:
                  type: integer
                  example: 2
                quantity:
                  type: integer
                  example: 2
                rentTime:
                  type: integer
                  description: Total rental time in days
                  example: 3
      responses:
        '201':
          description: Item added to cart successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "Item added to cart successfully" }
                  data:
                    $ref: '#/components/schemas/CartItem'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Product not found
...
  /api/orders:
    post:
      summary: Create a new order for a single product
      description: Accepts a cart item payload and creates a dedicated order.
      tags: [Orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Validation error (e.g., missing productId or totals)
        '401':
          description: Unauthorized
        '500':
          description: Server error

    get:
      summary: List the authenticated user's orders
      description: Returns all rental orders associated with the current session.
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  count: { type: integer, example: 3 }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
        '500':
          description: Server error
...
  /api/reports:
    post:
      summary: Create new report
      description: Submit a report about another user
      tags: [Reports]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reporting_user_id
                - reported_user_id
                - reason
              properties:
                reporting_user_id:
                  type: integer
                reported_user_id:
                  type: integer
                reporting_user_username:
                  type: string
                reported_user_username:
                  type: string
                reason:
                  type: string
                reason_extended:
                  type: string
                image:
                  type: string
                  format: uri
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Report'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Get all reports
      description: Retrieve paginated list of reports (Admin)
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_review, resolved, dismissed]
        - name: reporting_user_id
          in: query
          schema:
            type: integer
        - name: reported_user_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Reports retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      reports:
                        type: array
                        items:
                          $ref: '#/components/schemas/Report'
                      pagination:
                        $ref: '#/components/schemas/PaginationResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /api/reviews:
    post:
      summary: Submit a product review
      description: Creates a new review for a completed rental/order.
      tags: [Reviews]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreatePayload'
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "Review created successfully" }
                  data:
                    $ref: '#/components/schemas/Review'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Product not found
        '409':
          description: Review already exists for the provided order
...